+++ lib/sidekiq/web_helpers.rb
@@ -122,15 +122,7 @@ module Sidekiq
     end
 
     def redis_info
-      Sidekiq.redis do |conn|
-        # admin commands can't go through redis-namespace starting
-        # in redis-namespace 2.0
-        if conn.respond_to?(:namespace)
-          conn.redis.info
-        else
-          conn.info
-        end
-      end
+      Sidekiq.redis_info
     end
 
     def root_path
@@ -174,7 +166,7 @@ module Sidekiq
 
     def display_args(args, truncate_after_chars = 2000)
       args.map do |arg|
-        h(truncate(to_display(arg)))
+        h(truncate(to_display(arg), truncate_after_chars))
       end.join(", ")
     end